feat: tooltip and platform detection
